Creé la clase Contenido y la clase Documental dónde ahora Pelicula y Documental heredarán atributos y métodos de la clase  Contenido
